created definitions
pushed gathering of nodes and links into their own definition as well as the process of extracting the lengths from the inp file into its own definition. within the nodes and links are numerable global variables to be called on. the function get_nodes_and_links does not return any values as all relevant values are within the global variables.