developing branches definition
this update includes a while loop which gets and appends pollution quality per each node over time. still needs to get an array for each of the times at given dates.
The next major portion is development of a branches function which finds all paths from a start node to an end node and adds its lengths. currently length part doesn't work, however it is likely that being able to parse through the paths is functional.